local-deepseek/requirements.txt
accelerate==1.7.0
bitsandbytes==0.45.5
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
diskcache==5.6.3
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.9
MarkupSafe==3.0.2
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.7.1
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
typing_extensions==4.13.2
urllib3==2.4.0
